// src/app/models/empleado.model.ts
export class EmpleadoModel{
    documento: string;
    cargo: string;
    telefono: string;
    tipoDocumento: string;
    correo: string;
    fechaNacimiento: string;
    nombreCompleto: string;
    rh: string;
    direccion: string;
    municipio: string;
    departmento: string;
    foto: string;
}